+++ cdhweb/events/models.py
@@ -3,15 +3,11 @@
 from datetime import datetime
 
 import icalendar
-from django.contrib.auth.models import User
 from django.core.exceptions import ValidationError
 from django.db import models
 from django.urls import reverse
 from django.utils import timezone
 from django.utils.html import strip_tags
-from mezzanine.core.fields import FileField
-from mezzanine.core.models import Displayable, RichText
-from mezzanine.utils.models import AdminThumbMixin, upload_to
 from modelcluster.contrib.taggit import ClusterTaggableManager
 from modelcluster.fields import ParentalKey
 from modelcluster.models import ClusterableModel
@@ -24,8 +20,7 @@ from wagtail.core.models import Page, PageManager, PageQuerySet
 from wagtail.images.edit_handlers import ImageChooserPanel
 
 from cdhweb.pages.models import LinkPage, BasePage
-from cdhweb.resources.models import Attachment, ExcerptMixin
-from cdhweb.resources.utils import absolutize_url
+from cdhweb.people.models import Person
 
 
 class EventTypeManager(models.Manager):
@@ -97,140 +92,6 @@ class EventQuerySet(PageQuerySet):
         return self.filter(end_time__lt=today).order_by('-start_time')
 
 
-class OldEvent(Displayable, RichText, AdminThumbMixin, ExcerptMixin):
-    '''An event, such as a workshop, lecture, or conference.'''
-
-    # description = rich text field
-    # NOTE: do we want a sponsor field? or jest include in description?
-    sponsor = models.CharField(max_length=80, null=True, blank=True)
-    start_time = models.DateTimeField()
-    end_time = models.DateTimeField()
-    # all day flag todo
-    # all_day = models.BooleanField(default=False, blank=True)
-    location = models.ForeignKey(Location, null=True, blank=True,
-                                 on_delete=models.SET_NULL)
-    event_type = models.ForeignKey(EventType, on_delete=models.CASCADE)
-    speakers = models.ManyToManyField(User,
-                                      help_text='Guest lecturer(s) or Workshop leader(s)',
-                                      blank=True)
-
-    attendance = models.PositiveIntegerField(null=True, blank=True,
-                                             help_text='Total number of people who attended the event. (Internal only, for reporting purposes.)')
-
-    join_url = models.URLField(verbose_name="Join URL", null=True, blank=True,
-                               help_text='Join URL for virtual events, e.g. Zoom meetings.')
-
-    # TODO: include expected size? (required size?)
-    image = FileField(verbose_name="Image",
-                      upload_to=upload_to("events.image", "events"),
-                      format="Image", max_length=255, null=True, blank=True,
-                      help_text='Image for display on event detail page (optional)')
-
-    thumb = FileField(verbose_name="Thumbnail",
-                      upload_to=upload_to("events.thumb", "events/thumbnails"),
-                      format="Image", max_length=255, null=True, blank=True,
-                      help_text='Image for display on event card (optional)')
-
-    attachments = models.ManyToManyField(Attachment, blank=True)
-
-    tags = TaggableManager(blank=True)
-
-    # override manager for custom queryset filters
-    objects = EventQuerySet.as_manager()
-
-    admin_thumb_field = "thumb"
-    event_type.verbose_name = 'Type'
-
-    def __str__(self):
-        return ' - '.join([self.title, self.start_time.strftime('%b %d, %Y')])
-
-    class Meta:
-        ordering = ("start_time",)
-
-    def is_virtual(self):
-        '''If an event takes place in a virtual location, it is virtual'''
-        if self.location:
-            return self.location.is_virtual
-        return False
-    is_virtual.boolean = True
-    is_virtual.short_description = "Virtual"
-
-    def get_absolute_url(self):
-        '''event detail url on this site'''
-        # we don't have to worry about the various url config options
-        # that mezzanine has to support; just handle the url style we
-        # want to use locally
-        return reverse('event:detail', kwargs={
-            'year': self.start_time.year,
-            # force two-digit month
-            'month': '%02d' % self.start_time.month,
-            'slug': self.slug})
-
-    def full_url(self):
-        '''Absolute url, including site address'''
-        return absolutize_url(self.get_absolute_url())
-
-    def get_ical_url(self):
-        '''URL to download this event as ical'''
-        return reverse('event:ical', kwargs={
-            'year': self.start_time.year,
-            # force two-digit month
-            'month': '%02d' % self.start_time.month,
-            'slug': self.slug})
-
-    def when(self):
-        '''Event start/end date and time, formatted for display.
-        Removes leading zeros from hours and converts am/pm to lower case.'''
-
-        # NOTE: - in %-I is to remove leading zero
-        # (possibly platform specific?)
-
-        local_tz = timezone.get_default_timezone()
-        # convert dates to local timezone for display
-        local_start = self.start_time.astimezone(local_tz)
-        local_end = self.end_time.astimezone(local_tz)
-        start = ' '.join([local_start.strftime('%b %d'),
-                          local_start.strftime('%-I:%M')])
-        start_ampm = local_start.strftime('%p')
-        # include start am/pm if *different* from end
-        if start_ampm != local_end.strftime('%p'):
-            start += ' %s' % start_ampm.lower()
-
-        # include end month and day if *different* from start
-        end_pieces = []
-        if local_start.month != local_end.month:
-            end_pieces.append(local_end.strftime('%b %d'))
-        elif local_start.day != local_end.day:
-            end_pieces.append(local_end.strftime('%d'))
-        end_pieces.append(local_end.strftime('%-I:%M %p').lower())
-        end = ' '.join(end_pieces)
-
-        return ' – '.join([start, end])
-
-    def duration(self):
-        '''duration between start and end time as :class:`datetime.timedelta`'''
-        return self.end_time - self.start_time
-
-    def ical_event(self):
-        '''Return the current event as a :class:`icalendar.Event` for
-        inclusion in a :class:`icalendar.Calendar`'''
-        event = icalendar.Event()
-        # use absolute url for event id and in event content
-        absurl = self.full_url()
-        event.add('uid', absurl)
-        event.add('summary', self.title)
-        event.add('dtstart', self.start_time)
-        event.add('dtend', self.end_time)
-        if self.location:
-            if self.is_virtual() and self.join_url:
-                event.add('location', self.join_url)
-            else:
-                event.add('location', self.location.display_name)
-        event.add('description',
-                  '\n'.join([strip_tags(self.content), '', absurl]))
-        return event
-
-
 # custom manager for wagtail pages, see:
 # https://docs.wagtail.io/en/stable/topics/pages.html#custom-page-managers
 EventManager = PageManager.from_queryset(EventQuerySet)
@@ -245,11 +106,12 @@ class EventTag(TaggedItemBase):
 class Speaker(models.Model):
     """Relationship between Person and Event."""
     event = ParentalKey("events.Event", related_name="speakers")
-    person = models.ForeignKey("people.Person", on_delete=models.CASCADE)
+    person = models.ForeignKey(Person, on_delete=models.CASCADE)
     panels = [FieldPanel("person")]
 
     class Meta:
-        ordering = ("person__last_name", "person__first_name")
+        # can't sort on related fields due to modelcluster limitations
+        ordering = ("person",)
 
     def __str__(self) -> str:
         return "%s at %s" % (self.person, self.event)
@@ -274,6 +136,8 @@ class Event(BasePage, ClusterableModel):
     thumbnail = models.ForeignKey("wagtailimages.image", null=True, blank=True,
                                   on_delete=models.SET_NULL, related_name="+",
                                   help_text="Image for display on event card (optional)")
+    people = models.ManyToManyField(Person, through=Speaker,
+                                    related_name='events')
     tags = ClusterTaggableManager(through=EventTag, blank=True)
     # TODO attachments (#245)
 
@@ -304,9 +168,6 @@ class Event(BasePage, ClusterableModel):
 
     # custom manager/queryset logic
     objects = EventManager()
-
-    class Meta:
-        ordering = ("-start_time",)
 
     def __str__(self):
         return " - ".join([self.title, self.start_time.strftime("%b %d, %Y")])